mcp-server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=1.0.0
langchain-groq>=0.1.0
langchain-community>=0.0.30
faiss-cpu>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.10.0
requests>=2.31.0
httpx>=0.26.0
python-multipart>=0.0.6
